requirements.txt
git+https://github.com/amazon-science/chronos-forecasting.git
jupyterlab
chardet
plotly
nbformat
tqdm
tensorflow
matplotlib
torch
numpy
pandas
scikit-learn
u8darts[all]
dask-expr
dask
seaborn